[CASB-122624] add logging on corrupted session
This is to help aid during production outages related to a corrupted
session. We are using print logging because there is no logger available
here, so the logs will appear as lines in whatever log the starting
process outputs (typically the apache2 error log in
`/var/log/apache2/error.log` for ui portal).